ecommerce-microservices/order-service/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
gunicorn>=21.2
requests>=2.31
